GoRouter with state
