+++ lib/config/router/app_router.dart
@@ -1,20 +1,33 @@
 import 'package:cinemapedia/presentation/screens/screens.dart';
 import 'package:go_router/go_router.dart';
 
-final appRouter = GoRouter(initialLocation: '/', routes: [
-  GoRoute(
-    path: '/',
-    name: HomeScreen.name,
-    builder: (context, state) => const HomeScreen(),
-    routes: [
+final appRouter = GoRouter(
+  initialLocation: '/home/0',
+  routes: [
+
       GoRoute(
-        path: 'movie/:id',
-        name: MovieScreen.name,
+        path: '/home/:page',
+        name: HomeScreen.name,
         builder: (context, state) {
-          final movieId = state.pathParameters['id'] ?? 'no-id';
-          return MovieScreen(movieId: movieId);
+          final pageIndex = int.parse(state.pathParameters['page'] ?? '0');
+          return HomeScreen(pageIndex: pageIndex,);
         },
+        routes: [
+          GoRoute(
+            path: 'movie/:id',
+            name: MovieScreen.name,
+            builder: (context, state) {
+              final movieId = state.pathParameters['id'] ?? 'no-id';
+              return MovieScreen(movieId: movieId);
+            },
+          ),
+        ]
       ),
-    ]
-  )
-]);
+
+      GoRoute(
+        path: '/',
+        redirect: (_, __) => '/home/0',
+      ),
+
+  ]
+);
