+++ lib/presentation/widgets/shared/bottom_navigation.dart
@@ -1,12 +1,32 @@
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 class CustomBottomNavigation extends StatelessWidget {
-  const CustomBottomNavigation({super.key});
+  const CustomBottomNavigation({super.key, required this.currentIndex});
+
+  final int currentIndex;
+
+  void onItemTape(BuildContext context, int index) {
+    if (index == currentIndex) return;
+    switch (index) {
+      case 0:
+        context.go('/home/0');
+        break;
+      case 1:
+        context.go('/home/1');
+        break;
+      case 2:
+        context.go('/home/2');
+        break;
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return BottomNavigationBar(
+      onTap: (value) => onItemTape(context, value),
       elevation: 0,
+      currentIndex: currentIndex,
       items: const[
         BottomNavigationBarItem(
           icon: Icon(Icons.home_max),
